+++ lib/internal/crypto/hash.js
@@ -87,6 +87,11 @@ const maybeEmitDeprecationWarning = getDeprecationWarningEmitter(
   },
 );
 
+const emitHmacDigestDeprecation = getDeprecationWarningEmitter(
+  'DEP0206',
+  'Calling Hmac.digest() more than once is deprecated.',
+);
+
 function Hash(algorithm, options) {
   if (!new.target)
     return new Hash(algorithm, options);
@@ -183,6 +188,7 @@ Hmac.prototype.digest = function digest(outputEncoding) {
   const state = this[kState];
 
   if (state[kFinalized]) {
+    emitHmacDigestDeprecation();
     const buf = Buffer.from('');
     if (outputEncoding && outputEncoding !== 'buffer')
       return buf.toString(outputEncoding);
