+++ test/parallel/test-crypto-hmac.js
@@ -7,13 +7,6 @@ if (!common.hasCrypto) {
 const assert = require('assert');
 const crypto = require('crypto');
 
-{
-  const Hmac = crypto.Hmac;
-  const instance = crypto.Hmac('sha256', 'Node');
-  assert(instance instanceof Hmac, 'Hmac is expected to return a new instance' +
-                                   ' when called without `new`');
-}
-
 assert.throws(
   () => crypto.createHmac(null),
   {
@@ -411,41 +404,6 @@ assert.strictEqual(
   crypto.createHmac('sha256', 'w00t').digest('ucs2'),
   crypto.createHmac('sha256', 'w00t').digest().toString('ucs2'));
 
-// Check initialized -> uninitialized state transition after calling digest().
-{
-  const expected =
-      '\u0010\u0041\u0052\u00c5\u00bf\u00dc\u00a0\u007b\u00c6\u0033' +
-      '\u00ee\u00bd\u0046\u0019\u009f\u0002\u0055\u00c9\u00f4\u009d';
-  {
-    const h = crypto.createHmac('sha1', 'key').update('data');
-    assert.deepStrictEqual(h.digest('buffer'), Buffer.from(expected, 'latin1'));
-    assert.deepStrictEqual(h.digest('buffer'), Buffer.from(''));
-  }
-  {
-    const h = crypto.createHmac('sha1', 'key').update('data');
-    assert.strictEqual(h.digest('latin1'), expected);
-    assert.strictEqual(h.digest('latin1'), '');
-  }
-}
-
-// Check initialized -> uninitialized state transition after calling digest().
-// Calls to update() omitted intentionally.
-{
-  const expected =
-      '\u00f4\u002b\u00b0\u00ee\u00b0\u0018\u00eb\u00bd\u0045\u0097' +
-      '\u00ae\u0072\u0013\u0071\u001e\u00c6\u0007\u0060\u0084\u003f';
-  {
-    const h = crypto.createHmac('sha1', 'key');
-    assert.deepStrictEqual(h.digest('buffer'), Buffer.from(expected, 'latin1'));
-    assert.deepStrictEqual(h.digest('buffer'), Buffer.from(''));
-  }
-  {
-    const h = crypto.createHmac('sha1', 'key');
-    assert.strictEqual(h.digest('latin1'), expected);
-    assert.strictEqual(h.digest('latin1'), '');
-  }
-}
-
 {
   assert.throws(
     () => crypto.createHmac('sha7', 'key'),
@@ -459,14 +417,4 @@ assert.strictEqual(
     crypto.createHmac('sha256', buf).update('foo').digest(),
     crypto.createHmac('sha256', keyObject).update('foo').digest(),
   );
-}
-
-{
-  crypto.Hmac('sha256', 'Node');
-  common.expectWarning({
-    DeprecationWarning: [
-      ['crypto.Hmac constructor is deprecated.',
-       'DEP0181'],
-    ]
-  });
 }
